Separated mongodb connection logic from main function.
