+++ main.go
@@ -1,55 +1,13 @@
 package main
 
 import (
-	"context"
-	"fmt"
-	"log"
-	"time"
-
+	DbConnection "github.com/gemstack/jobs-api/dbconnection"
 	"github.com/gemstack/jobs-api/routes"
-	"go.mongodb.org/mongo-driver/bson"
-	"go.mongodb.org/mongo-driver/mongo"
-	"go.mongodb.org/mongo-driver/mongo/options"
-	"go.mongodb.org/mongo-driver/mongo/readpref"
 )
 
 func main() {
 
-	// Set connection options
-	clientOptions := options.Client().ApplyURI("mongodb+srv://jobsapireadnwrite1:AH6tUs5R2rQZcaN9@jobsapicluster0.jaetiqg.mongodb.net/")
-
-	// Connect to MongoDB
-	client, err := mongo.Connect(context.TODO(), clientOptions)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	//define timeout duration
-	// Pass a context with a timeout to tell a blocking function that it
-	// should abandon its work after the timeout elapses.
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	select {
-	case <-time.After(1 * time.Second):
-		fmt.Println("overslept")
-	case <-ctx.Done():
-		fmt.Println(ctx.Err()) // prints "context deadline exceeded"
-	}
-	defer client.Disconnect(ctx)
-
-	//ping the cluster from within app--if no err thrown, ping success!
-	err = client.Ping(ctx, readpref.Primary())
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	//print databases from connected cluster
-	databases, err := client.ListDatabaseNames(ctx, bson.M{})
-	if err != nil {
-		log.Fatal(err)
-	}
-	fmt.Println(databases)
+	DbConnection.DbConnection()
 
 	//r := gin.Default()
